requirements.txt
crewai
litellm
python-dotenv
pytest
